+++ src/main/java/ru/gb/ProductsConsoleApp/Cart.java
@@ -1,11 +1,13 @@
-package ru.gb.ProductsComsoleApp;
+package ru.gb.ProductsConsoleApp;
 
+import lombok.ToString;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
 import java.util.HashMap;
 import java.util.Map;
 
+@ToString
 @Component
 @Scope("prototype")
 public class Cart {
